refactor(billing): move some util functions from utils.py
And adjust the imports
